requirements.txt
google-play-scraper
pandas
google-generativeai
schedule
sentence-transformers
hdbscan
scikit-learn
